Moved validation and business logic to service layer

+++ src/main/java/Service/MessageService.java
@@ -1,7 +1,8 @@
 package Service;
 
-import java.util.List;
+import java.util.*;
 
+import DAO.AccountDAO;
 import DAO.MessageDAO;
 import Model.Message;
 
@@ -13,12 +14,28 @@ public class MessageService implements MessageServiceInt {
     }
 
     /**
+     * This method checks if a string is completely filled with whitespace.
+     * @param string - a String object.
+     * @return true if string is whitespace, false otherwise.
+     */
+    private boolean whiteSpaceChecker(String str) {
+        for (char ch : str.toCharArray()) {
+            if (!Character.isWhitespace(ch)) return false;
+        }
+        return true;
+    }
+
+    /**
      * This method creates a new message.
      * @param message - a message object.
      * @return newly created message.
      */
     public Message createMessage(Message message) {
-        return messageDAO.createMessage(message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch());
+        // message text must not be longer than 255 characters and cannot be empty, and account must exist
+        if ((message.getMessage_text().length() <= 255) && (!whiteSpaceChecker(message.getMessage_text())) && (AccountDAO.findAccountById(message.getPosted_by()) != null)) {
+            return messageDAO.createMessage(message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch());
+        }
+        return null;
     }
 
     /**
@@ -37,7 +54,11 @@ public class MessageService implements MessageServiceInt {
      * @return existing message.
      */
     public Message updateMessageById(int message_id, String updated_message_text) {
-        return messageDAO.updateMessageById(message_id, updated_message_text);
+        // updated text must not be longer than 255 characters and cannot be empty
+        if ((updated_message_text.length() <= 255) && (!whiteSpaceChecker(updated_message_text))) {
+            return messageDAO.updateMessageById(message_id, updated_message_text);
+        }
+        return null;
     }
 
     /**
@@ -46,7 +67,11 @@ public class MessageService implements MessageServiceInt {
      * @return newly deleted message.
      */
     public Message deleteMessageById(int id) {
-        return messageDAO.deleteMessageById(id);
+        // message must exist
+        if (messageDAO.getMessageById(id) != null) {
+            return messageDAO.deleteMessageById(id);
+        }
+        return null;
     }
 
     /**
